re-name stateful_tcp_world
